Fix for Dexcom G7
Expired Dexcom G7 transmitters usually stay 'paired' for a while, because the user doesn't delete them in the iOS bluetooth settings.
This changes makes sure that xDrip4iOS will not connected to those devices. It sets a timer waiting for an authentication response, if that's not received then it disconnects, forgets the devices and scans for a new one